examples/investigator/pyproject.toml
[project]
name = "investigator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Don Smith", email = "hi@dons.dev" }]
requires-python = ">=3.13,<3.14"
dependencies = [
  "langchain>=1.0.7",
  "langchain-anthropic>=1.0.4",
  "langgraph>=1.0.3",
  "langsmith>=0.4.43",
  "opentelemetry-api>=1.38.0",
  "opentelemetry-sdk>=1.38.0",
  "pydantic>=2.12.4",
  "pydantic-settings>=2.12.0",
  "pytest>=9.0.1",
  "python-dotenv>=1.2.1",
  "chromadb>=0.4.22",
  "neo4j>=5.0.0",
  "tree-sitter>=0.23.0",
  "tree-sitter-python>=0.23.0",
  "tree-sitter-javascript>=0.23.0",
  "chroma-mcp>=0.2.6",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["ruff"]
